requirements.txt
pandas
sentence-transformers 
faiss-cpu
pyarrow
accelerate
